Scout: Move utils to shared library
