+++ libs/shared-ui/src/lib/index.ts
@@ -9,5 +9,4 @@ export * from './icons';
 export * from './notifications';
 export * from './resources';
 export * from './sidebar';
-export * from './utils';
 export * from './web-socket';
