+++ libs/shared-ui/src/lib/resources/items/models.ts
@@ -1,7 +1,8 @@
-import { DEFAULT_SORT_ORDER, QueryParams, Recipe, SortOrder } from '@craftsmans-ledger/shared-ui';
 import { Expose } from 'class-transformer';
 import { nanoid } from 'nanoid';
-import { Resource } from '../models';
+import { DEFAULT_SORT_ORDER, Resource, SortOrder } from '../models';
+import { Recipe } from '../recipes';
+import { QueryParams } from '../../http';
 
 export class Item implements Resource {
     @Expose()
