+++ libs/shared-ui/test/mocks/api/handlers/item.handlers.ts
@@ -1,8 +1,9 @@
 import { delay, http, HttpHandler, HttpResponse } from 'msw';
-import { CreateItemData, Item, PaginatedResponse, serialize, tryCatch } from '../../../../src';
+import { CreateItemData, Item, PaginatedResponse } from '../../../../src';
 import { mockItemDB } from '../../db';
 import { baseApiUrl } from '../api.models';
 import { sendBadRequestExceptionResponse, sendExceptionResponse, sendNotFoundExceptionResponse } from './expections';
+import { serialize, tryCatch } from '@craftsmans-ledger/shared';
 
 const endPoint = '/items';
