+++ libs/shared/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/rxjs';
+export * from './lib/utils';
