+++ libs/shared/src/lib/utils/serialization.ts
